Rename module to avoid namespace conflicts
