+++ value.py
@@ -1,4 +1,4 @@
-from droppable.case import to_class
+from drops.case import to_class
 from filemanager import readinfo
 from constants import Constants
